+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## Unreleased
 
+## 0.11.0 (19 March 2026)
+
 - **Breaking:** `str(Document)` now returns a Python dict-like representation
-  instead of TOML text.  Use the new `as_toml()` method on `Document` and
-  `Item` for TOML serialisation.
+  instead of TOML text.
+  Use the new `as_toml()` method on `Document` and `Item` for TOML serialisation.
 - Support `__contains__` on `ScalarItem` (for strings).
 
 ## 0.10.0 (18 March 2026)